requirements.txt
Django>=4.2
djangorestframework>=3.14
pandas>=2.0
openpyxl>=3.1